+++ src/web/assets/editentry/src/EntryTypeSwitcher.js
@@ -42,7 +42,7 @@
                     $('#fields').html(response.fieldsHtml);
                     Craft.initUiElements($('#fields'));
                     Craft.appendHeadHtml(response.headHtml);
-                    Craft.appendFootHtml(response.bodyHtml);
+                    Craft.appendBodyHtml(response.bodyHtml);
 
                     // Update the slug generator with the new title input
                     if (typeof slugGenerator !== 'undefined') {