+++ src/web/assets/fieldsettings/src/fieldsettings.js
@@ -67,7 +67,7 @@
                 this.$container.html('').append($settings);
                 Craft.initUiElements(this.$container);
                 Craft.appendHeadHtml(response.data.headHtml);
-                Craft.appendFootHtml(response.data.footHtml);
+                Craft.appendBodyHtml(response.data.bodyHtml);
             }).catch(() => {
                 if (!this._ignoreFailedRequest) {
                     Craft.cp.displayError(Craft.t('app', 'A server error occurred.'));